# Small BufferWrapper optimizations