--- a/requirements-dev.txt
+++ b/requirements-dev.txt
@@ -14,6 +14,7 @@ pytest-cov>=6.2.1
 pytest-mock>=3.15.0
 pytest-verbose-parametrize>=1.7.0
 pytest-xdist>=3.8.0
+pytest-benchmark>=5.1.0
 pytest>=8.4.2
 responses>=0.25.8
 ruff>=0.12.12
@@ -24,3 +25,4 @@ ty==0.0.1a20
 sortedcontainers-stubs>=2.4.0
 kafka-python>=2.0.0
 sentry-protos==0.2.0
+