Profiling an Android app (total time 134.791) shows we spend ~45s (33%)
self-time in the combination of read_u8, read_u32, read_uleb128, and
_debug_group:

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
...
  4802055    4.445    0.000   25.891    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/android/dex/dex_base_utils.py:540(get_string)
  4843334    3.750    0.000    6.214    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:99(read_u16)
  4920612    5.074    0.000    8.503    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:317(read_string_with_length)
  4940286    0.760    0.000    0.767    0.000 {method 'decode' of 'bytes' objects}
  5226031    0.603    0.000    0.603    0.000 {method 'replace' of 'str' objects}
  6570153    0.674    0.000    0.674    0.000 {method 'append' of 'list' objects}
 13070542    4.154    0.000   21.984    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:174(read_uleb128)
 14244714   11.413    0.000   18.692    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:121(read_u32)
 14458538    1.250    0.000    1.250    0.000 /Users/chromy/.local/share/sentry-devenv/pythons/3.13.1/python/lib/python3.13/enum.py:795(<genexpr>)
 19121508    1.828    0.000    1.828    0.000 {built-in method _struct.unpack}
 22236754   13.114    0.000   22.275    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:77(read_u8)
 27626383    1.929    0.000    1.929    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:50(seek)
 47300341    2.830    0.000    2.830    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:25(__exit__)
 47300341    2.902    0.000    2.902    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:20(__enter__)
 47300341   11.009    0.000   13.919    0.000 /Users/chromy/work/launchpad/src/launchpad/parsers/buffer_wrapper.py:66(_debug_group)

Replace _debug_group with a decorator and a few other things
saves around 10% of processing time:

$ time launchpad size ~/Downloads/app-release.apk >/dev/null

________________________________________________________
Executed in   67.61 secs    fish           external
   usr time   54.03 secs    0.08 millis   54.03 secs
   sys time    7.13 secs    2.17 millis    7.13 secs

$ time launchpad size ~/Downloads/app-release.apk >/dev/null

________________________________________________________
Executed in   67.76 secs    fish           external
   usr time   54.00 secs    0.09 millis   54.00 secs
   sys time    7.25 secs    3.71 millis    7.24 secs

$ git co chromy/2025-10-07-opt
Switched to branch 'chromy/2025-10-07-opt'
$ time launchpad size ~/Downloads/app-release.apk >/dev/null

________________________________________________________
Executed in   60.82 secs    fish           external
   usr time   47.15 secs    0.06 millis   47.15 secs
   sys time    7.13 secs    1.38 millis    7.12 secs

$ time launchpad size ~/Downloads/app-release.apk >/dev/null

________________________________________________________
Executed in   60.77 secs    fish           external
   usr time   47.11 secs    0.06 millis   47.11 secs
   sys time    7.22 secs    1.46 millis    7.21 secs

Also add some benchmarks which were useful for investigating this.